backend/requirements.txt
fastapi
uvicorn
pillow
torch
transformers
accelerate
